+++ lib/diary.js
@@ -7,22 +7,22 @@ const api = 'https://lit-caverns-20261.herokuapp.com'
 // });
 
 
-// function getMeals() {
-//   $.ajax({
-//     method: 'GET',
-//     url: api + '/api/v1/meals',
-//     success: function (data) {
-//       populateMeals(data)
-//     }
-//   })
-// }
+function getMeals() {
+  $.ajax({
+    method: 'GET',
+    url: api + '/api/v1/meals',
+    success: function (data) {
+      populateMeals(data)
+    }
+  })
+}
 
 function populateMeals(data) {
   data.forEach(function (meal) {
     // debugger
-    // meal.foods.forEach(foodFunctions.foodRows);
-    shared.foodRows(meal.foods[0]);
+    meal.foods.forEach(shared.foodRows);
+    // shared.foodRows(meal.foods[0]);
   })
 }
 
-// module.exports = {getMeals}
+module.exports = {getMeals}
