+++ lib/foods.js
@@ -2,17 +2,17 @@ const shared = require('./shared.js')
 
 const $ = require('jQuery')
 const api = 'https://lit-caverns-20261.herokuapp.com'
-const foodsURL = "http://localhost:8080/foods.html"
 
-$(document).ready( function() {
-  if (document.URL == foodsURL) {
-    getFoods();
-  } else {
-    shared.getMeals();
-  }
-  setFoodListener();
-  filterFoods();
-});
+
+// $(document).ready( function() {
+//   if (document.URL == foodsURL) {
+//     getFoods();
+//   } else {
+//     shared.getMeals();
+//   }
+//   setFoodListener();
+//   filterFoods();
+// });
 
 function setFoodListener() {
   $('#food-form').on('submit', function(event) {
@@ -78,3 +78,5 @@ function validateInputs(name, calories) {
     }
   })
 }
+
+module.exports = {getFoods, setFoodListener, filterFoods}
