+++ lib/index.js
@@ -2,3 +2,4 @@ require('./shared')
 require('./foods')
 require('./diary')
 require('./foods.scss')
+require('./doc_ready')
