+++ lib/shared.js
@@ -1,6 +1,9 @@
 const $ = require('jQuery')
 const api = 'https://lit-caverns-20261.herokuapp.com'
 
+
+
+
 function foodRows(food) {
   let foodId = food.id
   $("#foods-header").after(
@@ -16,15 +19,15 @@ function foodRows(food) {
   focusListener(foodId);
 };
 
-function getMeals() {
-  $.ajax({
-    method: 'GET',
-    url: api + '/api/v1/meals',
-    success: function (data) {
-      populateMeals(data)
-    }
-  })
-}
+// function getMeals() {
+//   $.ajax({
+//     method: 'GET',
+//     url: api + '/api/v1/meals',
+//     success: function (data) {
+//       populateMeals(data)
+//     }
+//   })
+// }
 
 function iconListener(foodId) {
   $(`[data-id=${foodId}] .delete-icon`).on('click', deleteRow)
@@ -104,4 +107,4 @@ function deleteFood(id) {
   })
 }
 
-module.exports = {foodRows, getMeals, iconListener, focusListener, deleteRow}
+module.exports = {foodRows, iconListener, focusListener, deleteRow}
